Moved ng validation into its own file Moved common data manipulation and formatting into commons
