--- commons/RamUtils.ts
@@ -1,11 +1,0 @@
-/// <reference path="RamAPI.ts" />
-
-export class Helpers {
-    static applyMixins(derivedCtor: any, baseCtors: any[]) {
-        baseCtors.forEach(baseCtor => {
-            Object.getOwnPropertyNames(baseCtor.prototype).forEach(name => {
-                derivedCtor.prototype[name] = baseCtor.prototype[name];
-            });
-        });
-    }
-}
