--- commons/_CommonTypes.ts
@@ -1,2 +1,0 @@
-/// <reference path="RamAPI.ts" />
-/// <reference path="RamUtils.ts" />
