+++ frontend/typescript/commons/ram-ng-validators.ts
@@ -1,0 +1,13 @@
+import {Control} from '@angular/common';
+import {Utils} from '../../../commons/ram-utils';
+
+export class RAMNgValidators {
+    public static dateFormatValidator(c: Control) {
+        let v = c.value;
+        return v !== null && Utils.parseDate(v) === null ? {
+            validateDate: {
+                valid: false
+            }
+        } : null;
+    }
+}
