+++ frontend/typescript/components/commons/access-period/access-period.component.ts
@@ -1,5 +1,7 @@
 import {OnInit, Input, Output, EventEmitter, Component} from '@angular/core';
 import {Validators, ControlGroup, Control, FormBuilder, FORM_DIRECTIVES} from '@angular/common';
+import {Utils} from '../../../../../commons/ram-utils';
+import {RAMNgValidators} from '../../../commons/ram-ng-validators';
 
 @Component({
     selector: 'access-period',
@@ -21,9 +23,9 @@ export class AccessPeriodComponent implements OnInit {
 
     public ngOnInit() {
         this.form = this._fb.group({
-            'startDate': [Helper.dateToDDMMYYYY(this.data.startDate),
+            'startDate': [Utils.dateToDDMMYYYY(this.data.startDate),
                 Validators.compose([Validators.required, RAMNgValidators.dateFormatValidator])],
-            'endDate': [Helper.dateToDDMMYYYY(this.data.endDate),
+            'endDate': [Utils.dateToDDMMYYYY(this.data.endDate),
                 Validators.compose([RAMNgValidators.dateFormatValidator])],
             'noEndDate': [this.data.noEndDate]
         }, { validator: Validators.compose([this._isDateBefore('startDate', 'endDate')]) });
@@ -44,8 +46,8 @@ export class AccessPeriodComponent implements OnInit {
 
     private _isDateBefore = (startDateCtrlName: string, endDateCtrlName: string) => {
         return (cg: ControlGroup) => {
-            let startDate = Helper.parseDate((cg.controls[startDateCtrlName] as Control).value);
-            let endDate = Helper.parseDate((cg.controls[endDateCtrlName] as Control).value);
+            let startDate = Utils.parseDate((cg.controls[startDateCtrlName] as Control).value);
+            let endDate = Utils.parseDate((cg.controls[endDateCtrlName] as Control).value);
 
             return (startDate !== null && endDate !== null && startDate.getTime() > endDate.getTime()) ? {
                 isEndDateBeforeStartDate: {
@@ -54,57 +56,6 @@ export class AccessPeriodComponent implements OnInit {
             } : null;
         };
     }
-}
-
-export class RAMNgValidators {
-    public static dateFormatValidator(c: Control) {
-        let v = c.value;
-        return v !== null && Helper.parseDate(v) === null ? {
-            validateDate: {
-                valid: false
-            }
-        } : null;
-    }
-}
-
-export class Helper {
-
-    public static dateToDDMMYYYY(date: Date) {
-        if (date) {
-            return date.getDay() + '/' + date.getMonth() + '/' + date.getFullYear();
-        }
-    }
-
-    public static parseDate(dateString: string): Date {
-        if (dateString === null) {
-            return null;
-        }
-
-        if (!/^\d{1,2}\/\d{1,2}\/\d{4}$/.test(dateString)) {
-            return null;
-        }
-
-        let parts = dateString.split('/');
-        let day = parseInt(parts[0], 10);
-        let month = parseInt(parts[1], 10);
-        let year = parseInt(parts[2], 10);
-
-        if (year < 1000 || year > 3000 || month === 0 || month > 12) {
-            return null;
-        }
-
-        let monthLength = [31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
-
-        if (year % 400 === 0 || (year % 100 !== 0 && year % 4 === 0)) {
-            monthLength[1] = 29;
-        }
-
-        // Check the range of the day
-        if (day > 0 && day <= monthLength[month - 1]) {
-            return new Date(year, month, day);
-        }
-        return null;
-    };
 }
 
 export interface AccessPeriodComponentData {
